add color hue change

+++ src/wid_main_menu.cpp
@@ -231,9 +231,12 @@ color color_change_hue(const color &in, const float fHue)
       {1.0f / 3.0f * (1.0f - cosA) - sqrtf(1.0f / 3.0f) * sinA,
        1.0f / 3.0f * (1.0f - cosA) + sqrtf(1.0f / 3.0f) * sinA, cosA + 1.0f / 3.0f * (1.0f - cosA)}};
   // Use the rotation matrix to convert the RGB directly
-  out.r = clamp(in.r * matrix[ 0 ][ 0 ] + in.g * matrix[ 0 ][ 1 ] + in.b * matrix[ 0 ][ 2 ]);
-  out.g = clamp(in.r * matrix[ 1 ][ 0 ] + in.g * matrix[ 1 ][ 1 ] + in.b * matrix[ 1 ][ 2 ]);
-  out.b = clamp(in.r * matrix[ 2 ][ 0 ] + in.g * matrix[ 2 ][ 1 ] + in.b * matrix[ 2 ][ 2 ]);
+  out.r = clamp(((float) in.r) * matrix[ 0 ][ 0 ] + ((float) in.g) * matrix[ 0 ][ 1 ] +
+                ((float) in.b) * matrix[ 0 ][ 2 ]);
+  out.g = clamp(((float) in.r) * matrix[ 1 ][ 0 ] + ((float) in.g) * matrix[ 1 ][ 1 ] +
+                ((float) in.b) * matrix[ 1 ][ 2 ]);
+  out.b = clamp(((float) in.r) * matrix[ 2 ][ 0 ] + ((float) in.g) * matrix[ 2 ][ 1 ] +
+                ((float) in.b) * matrix[ 2 ][ 2 ]);
   return out;
 }
